Move ProcessGeofeed to 'verify' sub-package
